refactor extension background into modules
