+++ extension/background/state.js
@@ -1,0 +1,16 @@
+export const state = {
+  buffer: [],
+  lastFlush: Date.now(),
+  seq: 0,
+  dropBodies: false,
+  runMeta: null,
+  pendingResponses: new Map(),
+  activeDebuggers: new Map(),
+  counters: {
+    http_req: 0,
+    http_res: 0,
+    ws_open: 0,
+    ws_frames: 0,
+    dropped_bodies: 0,
+  },
+};
