+++ extension/manifest.json
@@ -5,7 +5,8 @@
   "description": "Capture network activity via Chrome DevTools Protocol",
   "permissions": ["debugger", "tabs", "activeTab", "downloads", "storage"],
   "background": {
-    "service_worker": "background.js"
+    "service_worker": "background/main.js",
+    "type": "module"
   },
   "action": {
     "default_popup": "popup.html"
